ner_dataset_builder/requirements.txt
pyyaml>=6.0.1
torch>=2.1.0
tqdm>=4.66.0
transformers>=4.40.0
